break out string parsing functions
